update 3.5.0 fix authority dashboards

+++ car_vendor/views.py
@@ -1,18 +1,32 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
-from .forms import CarVendorProfileFrom,CarForm
-
-# Create your views here.
+from management.views import allow_access
+from .forms import CarVendorProfileFrom, CarForm
 from car_vendor.models import Car, CarVendorProfile
 
 
 @login_required(login_url="/login/")
 def index(request):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
+    try:
+        profile = CarVendorProfile.objects.get(user=request.user)
+    except CarVendorProfile.DoesNotExist:
+        return redirect('car_vendor_profile')
     cars = Car.objects.filter(user=request.user)
-    return render(request, 'car_vendor/index.html',{'cars':cars})
+    return render(request, 'car_vendor/index.html', {'cars': cars, 'access_level': access_level})
+
 
 @login_required(login_url="/login/")
 def add_car(request):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
+    try:
+        profile = CarVendorProfile.objects.get(user=request.user)
+    except CarVendorProfile.DoesNotExist:
+        return redirect('car_vendor_profile')
     if request.method == "POST":
         form = CarForm(request.POST, request.FILES)
         if form.is_valid():
@@ -25,27 +39,50 @@ def add_car(request):
             msg = 'Form is not valid'
     else:
         form = CarForm()
-    return render(request, 'car_vendor/add_car.html',{'form':form})
+    return render(request, 'car_vendor/add_car.html', {'form': form, 'access_level': access_level})
+
 
 @login_required(login_url="/login/")
-def delete_car(request,id):
+def delete_car(request, id):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
+    try:
+        profile = CarVendorProfile.objects.get(user=request.user)
+    except CarVendorProfile.DoesNotExist:
+        return redirect('car_vendor_profile')
     car = Car.objects.get(id=id)
-    return render(request, 'car_vendor/delete_car.html',{'car':car})
+    return render(request, 'car_vendor/delete_car.html', {'car': car, 'access_level': access_level})
+
 
 @login_required(login_url="/login/")
-def delete_car_check(request,id,check):
+def delete_car_check(request, id, check):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
+    try:
+        profile = CarVendorProfile.objects.get(user=request.user)
+    except CarVendorProfile.DoesNotExist:
+        return redirect('car_vendor_profile')
     car = Car.objects.get(id=id)
     if check:
         car.delete()
         return redirect('car_vendor_home')
-    return render(request, 'car_vendor/delete_car.html',{'car':car})
+    return render(request, 'car_vendor/delete_car.html', {'car': car, 'access_level': access_level})
 
 
 @login_required(login_url="/login/")
-def update_car(request,id):
+def update_car(request, id):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
+    try:
+        profile = CarVendorProfile.objects.get(user=request.user)
+    except CarVendorProfile.DoesNotExist:
+        return redirect('car_vendor_profile')
     car = Car.objects.get(id=id)
     if request.method == "POST":
-        form = CarForm(request.POST, request.FILES,instance=car)
+        form = CarForm(request.POST, request.FILES, instance=car)
         if form.is_valid():
             shared_data = form.save(commit=False)
             shared_data.user = request.user
@@ -56,10 +93,14 @@ def update_car(request,id):
             msg = 'Form is not valid'
     else:
         form = CarForm(instance=car)
-    return render(request, 'car_vendor/add_car.html',{'form':form})
+    return render(request, 'car_vendor/add_car.html', {'form': form, 'access_level': access_level})
+
 
 @login_required(login_url="/login/")
 def car_vendor_profile(request):
+    access_level = allow_access(request, ['car_vendor'])
+    if not access_level:
+        return redirect('role_elevation')
     try:
         profile = CarVendorProfile.objects.get(user=request.user)
     except CarVendorProfile.DoesNotExist:
@@ -77,4 +118,4 @@ def car_vendor_profile(request):
     else:
 
         form = CarVendorProfileFrom(instance=profile)
-    return render(request, 'car_vendor/car_vendor_profile.html',{'form':form})
+    return render(request, 'car_vendor/car_vendor_profile.html', {'form': form, 'profile': profile, 'access_level': access_level})
